Refactor authentication
Previously authentication seems to happen in a few files, where it would be nice if it was all in one place.

There was one usage in model.tsx but that file/api endpoint didn't seem to be used anymore.
