--- pages/api/models.ts
@@ -1,76 +1,0 @@
-import { OPENAI_API_HOST, OPENAI_API_TYPE, OPENAI_API_VERSION, OPENAI_ORGANIZATION, AZURE_APIM } from '@/utils/app/const';
-import { OpenAIModel, OpenAIModelID, OpenAIModels } from '@/types/openai';
-import { getAuthToken } from '@/utils/lib/azure';
-import type { NextApiRequest, NextApiResponse } from 'next';
-
-
-const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  try {
-    const { key } = req.body as {
-      key: string;
-    };
-
-    let url = `${OPENAI_API_HOST}/v1/models`;
-    if (OPENAI_API_TYPE === 'azure') {
-      url = `${OPENAI_API_HOST}/openai/deployments?api-version=${OPENAI_API_VERSION}`;
-    }
-
-    let token = await getAuthToken();
-    // TODO: Seems to currently not be working. Also doesn't seem to be used right now.
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        ...(OPENAI_API_TYPE === 'openai' && {
-          Authorization: `Bearer ${key ? key : process.env.OPENAI_API_KEY}`
-        }),
-        ...(OPENAI_API_TYPE === 'azure' && process.env.AZURE_USE_MANAGED_IDENTITY=="false" && {
-          'api-key': `${key ? key : process.env.OPENAI_API_KEY}`
-        }),
-        ...(OPENAI_API_TYPE === 'azure' && process.env.AZURE_USE_MANAGED_IDENTITY=="true" && {
-          Authorization: `Bearer ${token.token}`
-        }),
-        ...((OPENAI_API_TYPE === 'openai' && OPENAI_ORGANIZATION) && {
-          'OpenAI-Organization': OPENAI_ORGANIZATION,
-        }),
-        ...((AZURE_APIM) && {
-          'Ocp-Apim-Subscription-Key': process.env.AZURE_APIM_KEY
-        })
-      },
-    });
-
-    if (response.status === 401) {
-      res.status(500).json({ error: 'Unauthorized' });
-      return;
-    } else if (response.status !== 200) {
-      console.error(
-        `OpenAI API returned an error ${response.status
-        }: ${await response.text()}`,
-      );
-      throw new Error('OpenAI API returned an error');
-    }
-
-    const json = await response.json();
-
-    const models: OpenAIModel[] = json.data
-      .map((model: any) => {
-        const model_name = (OPENAI_API_TYPE === 'azure') ? model.model : model.id;
-        for (const [key, value] of Object.entries(OpenAIModelID)) {
-          if (value === model_name) {
-            return {
-              id: model.id,
-              name: OpenAIModels[value].name,
-            };
-          }
-        }
-      })
-      .filter(Boolean);
-    console.log("loaded models");
-    res.status(200).json(models);
-  } catch (error) {
-    console.error(error);
-    res.status(500).send('Error');
-  }
-};
-
-export default handler;
