+++ pages/home/index.ts
@@ -1,3 +1,1 @@
-//import { DefaultAzureCredential } from "@azure/identity";
 export { default, getServerSideProps } from './home';
-//export let azcredential = new DefaultAzureCredential();
